requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
aiomysql==0.2.0
pymysql==1.1.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# File Upload & Image Processing
cloudinary==1.38.0
python-multipart==0.0.6
pillow==10.2.0

# Templates
jinja2==3.1.3

# WSGI/ASGI Compatibility (for WSGI-only hosts)
asgiref==3.7.2

